+++ dotnet/Services/CybersourcePaymentService.cs
@@ -66,6 +66,8 @@ namespace Cybersource.Services
                     orderSuffix = merchantSettings.OrderSuffix.Trim();
                 }
 
+                string referenceNumber = await _vtexApiService.GetOrderId(createPaymentRequest.Reference);
+
                 if (createPaymentRequest.MerchantSettings != null)
                 {
                     foreach (MerchantSetting merchantSetting in createPaymentRequest.MerchantSettings)
@@ -122,7 +124,7 @@ namespace Cybersource.Services
                     },
                     clientReferenceInformation = new ClientReferenceInformation
                     {
-                        code = $"{createPaymentRequest.OrderId}{orderSuffix}", // Use Reference to have a consistent number with Fraud?
+                        code = $"{referenceNumber}{orderSuffix}", // Use Reference to have a consistent number with Fraud?
                         //transactionId = createPaymentRequest.TransactionId,
                         applicationName = $"{_context.Vtex.App.Vendor}.{_context.Vtex.App.Name}",
                         applicationVersion = _context.Vtex.App.Version,
@@ -826,11 +828,19 @@ namespace Cybersource.Services
                     return cancelPaymentResponse;
                 }
 
+                string referenceNumber = await _vtexApiService.GetOrderId(paymentData.OrderId);
+                string orderSuffix = string.Empty;
+                MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
+                {
+                    orderSuffix = merchantSettings.OrderSuffix.Trim();
+                }
+
                 Payments payment = new Payments
                 {
                     clientReferenceInformation = new ClientReferenceInformation
                     {
-                        code = await _vtexApiService.GetOrderId(paymentData.OrderId),
+                        code = $"{referenceNumber}{orderSuffix}",
                         applicationName = _context.Vtex.App.Name,
                         applicationVersion = _context.Vtex.App.Version,
                         applicationUser = _context.Vtex.App.Vendor,
@@ -912,12 +922,18 @@ namespace Cybersource.Services
                 }
 
                 string referenceNumber = await _vtexApiService.GetOrderId(paymentData.OrderId);
+                string orderSuffix = string.Empty;
+                MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
+                {
+                    orderSuffix = merchantSettings.OrderSuffix.Trim();
+                }
 
                 Payments payment = new Payments
                 {
                     clientReferenceInformation = new ClientReferenceInformation
                     {
-                        code = referenceNumber,
+                        code = $"{referenceNumber}{orderSuffix}",
                         applicationName = _context.Vtex.App.Name,
                         applicationVersion = _context.Vtex.App.Version,
                         applicationUser = _context.Vtex.App.Vendor,
@@ -1010,11 +1026,19 @@ namespace Cybersource.Services
             try
             {
                 PaymentData paymentData = await _cybersourceRepository.GetPaymentData(refundPaymentRequest.PaymentId);
+                string referenceNumber = await _vtexApiService.GetOrderId(refundPaymentRequest.PaymentId);
+                string orderSuffix = string.Empty;
+                MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
+                {
+                    orderSuffix = merchantSettings.OrderSuffix.Trim();
+                }
+
                 Payments payment = new Payments
                 {
                     clientReferenceInformation = new ClientReferenceInformation
                     {
-                        code = await _vtexApiService.GetOrderId(refundPaymentRequest.PaymentId),
+                        code = $"{referenceNumber}{orderSuffix}",
                         applicationName = _context.Vtex.App.Name,
                         applicationVersion = _context.Vtex.App.Version,
                         applicationUser = _context.Vtex.App.Vendor,
@@ -1276,7 +1300,7 @@ namespace Cybersource.Services
             if (paymentData != null && paymentData.CreatePaymentResponse != null)
             {
                 _context.Vtex.Logger.Debug("SetupPayerAuth", null, "Loaded PaymentData", new[] { ("orderId", createPaymentRequest.OrderId), ("paymentId", createPaymentRequest.PaymentId), ("createPaymentRequest", JsonConvert.SerializeObject(createPaymentRequest)), ("paymentData", JsonConvert.SerializeObject(paymentData)) });
-                await _vtexApiService.ProcessConversions();
+                //await _vtexApiService.ProcessConversions();
                 return paymentData.CreatePaymentResponse;
             }